fix: add existing channels
On some occasions, when a pre-release branch is created
from a commit on another pre-release branch and pushed
without any new commits, it's channel information will overwrite the
existing pre-release channel information. This change checks to
see if there's an existing note on the head commit and
adds the new channel to that existing note if there is one.